extract base class for isracard and amex
